+++ web-api/routers/__init__.py
@@ -15,6 +15,7 @@ from .physics import router as physics_router
 from .docs import router as docs_router
 from .tec import router as tec_router
 from .tid import router as tid_router
+from .dashboard import router as dashboard_router
 
 __all__ = [
     'health_router',
@@ -29,5 +30,6 @@ __all__ = [
     'physics_router',
     'docs_router',
     'tec_router',
-    'tid_router'
+    'tid_router',
+    'dashboard_router'
 ]